requirements.txt
streamlit
sentence-transformers
numpy
pandas
streamlit
tensorflow
transformers
bert-extractive-summarizer
